+++ apps/extension/src/core/domains/ethereum/handler.extension.ts
@@ -1,5 +1,4 @@
 import { DEBUG } from "@core/constants"
-import { chaindataProvider } from "@core/domains/chaindata"
 import {
   AnyEthRequestChainId,
   CustomEvmNetwork,
@@ -18,6 +17,7 @@ import { EthRequestArguments, EthRequestSignatures } from "@core/injectEth/types
 import { talismanAnalytics } from "@core/libs/Analytics"
 import { ExtensionHandler } from "@core/libs/Handler"
 import { watchEthereumTransaction } from "@core/notifications"
+import { chaindataProvider } from "@core/rpcs/chaindata"
 import { MessageTypes, RequestTypes, ResponseType } from "@core/types"
 import { Port, RequestIdOnly } from "@core/types/base"
 import { getPrivateKey } from "@core/util/getPrivateKey"
@@ -290,7 +290,7 @@ export class EthHandler extends ExtensionHandler {
     const networkId = parseInt(network.chainId, 16).toString()
     const newToken: CustomEvmNativeToken | null = network.nativeCurrency
       ? {
-          id: `${networkId}-native-${network.nativeCurrency.symbol}`.toLowerCase(),
+          id: `${networkId}-evm-native-${network.nativeCurrency.symbol}`.toLowerCase(),
           type: "evm-native",
           isTestnet: false,
           symbol: network.nativeCurrency.symbol,