+++ apps/extension/src/ui/index.tsx
@@ -3,6 +3,7 @@ import "@talisman/styles/styles.css"
 import { initSentry } from "@core/config/sentry"
 import * as Sentry from "@sentry/react"
 import { ErrorBoundary } from "@talisman/components/ErrorBoundary"
+import { ErrorBoundaryDatabaseMigration } from "@talisman/components/ErrorBoundaryDatabaseMigration"
 import { NotificationsContainer } from "@talisman/components/Notifications/NotificationsContainer"
 import ThemeProvider from "@talisman/theme"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
@@ -30,20 +31,22 @@ export const renderTalisman = (app: any) => {
     <React.StrictMode>
       <ThemeProvider>
         <ErrorBoundary>
-          <QueryClientProvider client={queryClient}>
-            <FeaturesProvider>
-              <SettingsProvider>
-                <AddressBookProvider>
-                  <AppStateProvider>
-                    <DbCacheProvider>
-                      <HashRouter>{app}</HashRouter>
-                    </DbCacheProvider>
-                  </AppStateProvider>
-                </AddressBookProvider>
-              </SettingsProvider>
-            </FeaturesProvider>
-            <NotificationsContainer />
-          </QueryClientProvider>
+          <ErrorBoundaryDatabaseMigration>
+            <QueryClientProvider client={queryClient}>
+              <FeaturesProvider>
+                <SettingsProvider>
+                  <AddressBookProvider>
+                    <AppStateProvider>
+                      <DbCacheProvider>
+                        <HashRouter>{app}</HashRouter>
+                      </DbCacheProvider>
+                    </AppStateProvider>
+                  </AddressBookProvider>
+                </SettingsProvider>
+              </FeaturesProvider>
+              <NotificationsContainer />
+            </QueryClientProvider>
+          </ErrorBoundaryDatabaseMigration>
         </ErrorBoundary>
       </ThemeProvider>
     </React.StrictMode>